10.1C_Iteration_8/10.1C_Iteration_8/Inventory.cs
namespace _10._1C_Iteration_8
{
    public class Inventory
    {
        private List<Item> _items = new List<Item>();

        public Inventory()
        {

        }

        public bool HasItem(string id)
        {
            foreach (Item item in _items)
            {
                if (item.AreYou(id))
                {
                    return true;
                }
            }
            return false;
        }

        public void Put(Item item)
        {
            _items.Add(item);
        }

        public Item Take(string id)
        {
            Item item = Fetch(id);
            if (item != null)
                _items.Remove(item);

            return item;
        }

        public Item Fetch(string id)
        {
            foreach (Item item in _items)
            {
                if (item.AreYou(id))
                {
                    return item;
                }
            }
            return null;
        }

        public string ItemList
        {
            get
            {
                string result = "";
                foreach (Item item in _items)
                {
                    result += "\t" + item.Name + "\n";
                }
                return result;
            }
        }
    }
}
